+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# 1.0.1
+
+- Support additional bundler arguments to be passed when using docker (#2)
+
 # 1.0.0
 
-Feature parity with `serverless-wsgi`.
+- Feature parity with `serverless-wsgi`.
 
 # 0.0.1
 
-Initial release.
+- Initial release.